+++ stubs/app/client/components/Banners.ts
@@ -1,10 +1,2 @@
-import {AppModel} from 'app/client/models/AppModel';
-import {DocPageModel} from 'app/client/models/DocPageModel';
+export { buildHomeBanners, buildDocumentBanners } from 'app/client/components/CoreBanners';
 
-export function buildHomeBanners(_app: AppModel) {
-  return null;
-}
-
-export function buildDocumentBanners(_docPageModel: DocPageModel) {
-  return null;
-}
