+++ stubs/app/client/ui/ActivationPage.ts
@@ -1,12 +1,7 @@
-import {AppModel} from 'app/client/models/AppModel';
-import {Disposable} from 'grainjs';
+import {
+  DefaultActivationPage, IActivationPageCreator
+} from "app/client/ui/DefaultActivationPage";
 
-export class ActivationPage extends Disposable {
-  constructor(_appModel: AppModel) {
-    super();
-  }
-
-  public buildDom() {
-    return null;
-  }
+export function getActivationPage(): IActivationPageCreator {
+  return DefaultActivationPage;
 }
